底盘、云台基本调通、恢复了部分保护
底盘、云台基本调通、恢复了部分保护

+++ ROBOT/ANALYSIS/can2_analysis.c
@@ -72,7 +72,7 @@ void CAN2_Feedback_Analysis(CanRxMsg *rx_message)
 函数返回值： 无
 描述：将数据存入TxMessage结构体再由CAN_Transmit发送
 ****************************************************/
-void CAN2_Chassis_SendMsg(int motor_201,int motor_202,int motor_203,int motor_204)
+void CAN2_Chassis_SendMsg(int16_t motor_201,int16_t motor_202,int16_t motor_203,int16_t motor_204)
 {	
 		CanTxMsg TxMessage;
 	  TxMessage.StdId = 0x200;      //帧ID为传入参数的CAN_ID
